move projects.py to the widgets module.
