--- qwatson/managers/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright © 2018 Jean-Sébastien Gosselin
-# https://github.com/jnsebgosselin/qwatson
-#
-# This file is part of QWatson.
-# Licensed under the terms of the GNU General Public License.
